agent/requirements.txt
langgraph
langchain
langchain-community
langchain-google-genai
langchain-mongodb
langchain-voyageai
pymongo
voyageai
python-dotenv
fastapi
uvicorn
pydantic
pypdf
beautifulsoup4
lxml
